pyproject.toml
[tool.poetry]
name = "fluss-server"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
balder = { version = "0.1.22", extras = ["subscription"] }
channels-redis = "^4.0.0"
omegaconf = "^2.2.3"
rich = "^12.6.0"
django-taggit = "^3.0.0"
lok = "^0.2.0"
django-health-check = "^3.17.0"
django-probes = "^1.7.0"
django-guardian = "^2.4.0"
psycopg2 = "^2.9.4"
namegenerator = "^1.0.6"
django-storages = "^1.13.1"
boto3 = "^1.24.89"
pillow = "^9.2.0"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
